Make the client thread-safe
The RavenClient is not marked as being thread-safe, but a single
instance is used within the target. In addition, we mutate the Logger
property to match the logger of the current event, meaning there could
be a race condition between two threads changing the Logger and sending
their message, causing messages to come from the wrong logger.

If this is a performance bottleneck, then perhaps a client could be
created per logger (assuming that the rest of the client is
thread-safe).

Parsing the DSN can fail with exceptions, therefore, this is preferable
to fail at the point of creating the client, rather than which trying
to de-serialize the config.

A specific use case for this delayed failure is when using NLog with
autoReload enabled, it will now only fail when trying to write
messages, but can then still be re-configured as long as the lazy has
not been successfully initialized.